Update
some work on shooting
beam debug
weaponfire displaying

+++ logic/GridElement.cpp
@@ -42,7 +42,7 @@ bool tower_defense::GridElement::setTurret(Turret *t) {
 }
 
 tower_defense::GridElement *tower_defense::GridElement::getUpNeighbour() const {
-    int y = this->location.getY() - 1;
+    double y = this->location.getY() - 1;
 
     GridElement* temp = grid.getElement(Point(this->location.getX(), y));
 
@@ -50,7 +50,7 @@ tower_defense::GridElement *tower_defense::GridElement::getUpNeighbour() const {
 }
 
 tower_defense::GridElement *tower_defense::GridElement::getDownNeighbour() const {
-    int y = this->location.getY() + 1;
+    double y = this->location.getY() + 1;
 
     GridElement* temp = grid.getElement(Point(this->location.getX(), y));
 
@@ -58,7 +58,7 @@ tower_defense::GridElement *tower_defense::GridElement::getDownNeighbour() const
 }
 
 tower_defense::GridElement *tower_defense::GridElement::getLeftNeighbour() const {
-    int x = this->location.getX() - 1;
+    double x = this->location.getX() - 1;
 
     GridElement* temp = grid.getElement(Point(x, this->location.getY()));
 
@@ -66,7 +66,7 @@ tower_defense::GridElement *tower_defense::GridElement::getLeftNeighbour() const
 }
 
 tower_defense::GridElement *tower_defense::GridElement::getRightNeighbour() const {
-    int x = this->location.getX() + 1;
+    double x = this->location.getX() + 1;
 
     GridElement* temp = grid.getElement(Point(x, this->location.getY()));
